Thread Sleeping
The thread wasn't alive enough for the audio to play, and the stream was borrow-checked away before it could do it's thing too. Now that's solved! I love this so much, been wanting to do this for a really long time. Now I want to try reading files and making sound out of them! Was going to do that in JS with Node, but this is much more ergonomic come to think of it! It might even be able to work in the browser with WASM too :D