WAV File Writing (v2)
This one seems to work properly! Had to explain that chunking seemed to be the issue, and this fixes the playback now. I wonder if this makes it so I won't be able to add custom user input to customize the sound on the fly though. I really want to add that.