The Sounds of GarageBand
Here's an example of what this can do. It lets you hear the sound of a binary file. Say like the executable for GarageBand on macOS, you can play it as audio! This concept is inspired from "The Sounds of Microsoft Paint" on YouTube. Now I want to make a whole GUI app out of this!